% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{annotate_kin_matrix}
\alias{annotate_kin_matrix}
\title{Annotate kinship matrix with hover text for mating pair comparisons}
\usage{
annotate_kin_matrix(matrix, pedigree.living, studbook)
}
\arguments{
\item{matrix}{A kinship matrix (usually filtered to living individuals)}

\item{pedigree.living}{A pedigree object from \code{pedtools::ped()} for the living population}

\item{studbook}{A tibble containing metadata for individuals}
}
\value{
A matrix of hover text strings matching dimensions of \code{matrix}
}
\description{
Annotate kinship matrix with hover text for mating pair comparisons
}
