% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{calculate_age}
\alias{calculate_age}
\title{Calculate age in years}
\usage{
calculate_age(birth, date)
}
\arguments{
\item{birth}{Date of birth}

\item{date}{Reference date}
}
\value{
Integer age in years
}
\description{
Calculate age in years
}
