% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/life_tables.R
\name{cohort_lifeTab}
\alias{cohort_lifeTab}
\title{Generate life table from studbook and timeline over cohorts}
\usage{
cohort_lifeTab(timeline, studbook, minYear, maxYear, span, maxAge)
}
\arguments{
\item{timeline}{Timeline tibble}

\item{studbook}{Studbook tibble}

\item{minYear}{First cohort year}

\item{maxYear}{Last cohort year}

\item{span}{Years per cohort}

\item{maxAge}{Maximum age}
}
\value{
A tibble with cohort life table summary
}
\description{
Generate life table from studbook and timeline over cohorts
}
