% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{count_births}
\alias{count_births}
\title{Count births and track individual presence over time}
\usage{
count_births(timeline, studbook)
}
\arguments{
\item{timeline}{A tibble with at least \code{ID}, \code{Date}, and \code{TypeEvent} columns}

\item{studbook}{A studbook tibble with birth/death info}
}
\value{
A tibble of births per ID by year
}
\description{
Count births and track individual presence over time
}
