% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigree_stats.R
\name{F_vector}
\alias{F_vector}
\title{Calculate inbreeding vector from a pedigree}
\usage{
F_vector(pedigree.living, studbook)
}
\arguments{
\item{pedigree.living}{A pedigree object}

\item{studbook}{Studbook tibble}
}
\value{
Numeric vector of inbreeding coefficients
}
\description{
Calculate inbreeding vector from a pedigree
}
