% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigree_stats.R
\name{family_history}
\alias{family_history}
\title{Summarize family-level stats for each individual}
\usage{
family_history(pedigree.living, studbook)
}
\arguments{
\item{pedigree.living}{Pedigree object for living individuals}

\item{studbook}{Studbook tibble}
}
\value{
A tibble with family metrics and metadata
}
\description{
Summarize family-level stats for each individual
}
