% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigree_stats.R
\name{founder_contributions}
\alias{founder_contributions}
\title{Calculate founder contributions to living population}
\usage{
founder_contributions(studbook, pedigree.living)
}
\arguments{
\item{studbook}{Studbook tibble}

\item{pedigree.living}{Pedigree for living individuals}
}
\value{
A named numeric vector of founder contribution weights
}
\description{
Calculate founder contributions to living population
}
