% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigree_stats.R
\name{founder_reps}
\alias{founder_reps}
\title{Return list of founder IDs represented in living population}
\usage{
founder_reps(pedigree, pedigree.living, studbook)
}
\arguments{
\item{pedigree}{Full pedigree object}

\item{pedigree.living}{Pedigree for living individuals}

\item{studbook}{Studbook tibble}
}
\value{
A character vector of founder IDs
}
\description{
Return list of founder IDs represented in living population
}
