% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigree_stats.R
\name{founder_summary}
\alias{founder_summary}
\title{Summarize metadata and contributions of founders}
\usage{
founder_summary(pedigree, pedigree.living, studbook)
}
\arguments{
\item{pedigree}{Full pedigree object}

\item{pedigree.living}{Pedigree object filtered to living individuals}

\item{studbook}{Tibble containing individual metadata}
}
\value{
A tibble summarizing contribution and attributes for each founder
}
\description{
Summarize metadata and contributions of founders
}
