% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigree_stats.R
\name{gen_numbers}
\alias{gen_numbers}
\title{Return number of generations for individuals}
\usage{
gen_numbers(pedigree.living)
}
\arguments{
\item{pedigree.living}{Pedigree object}
}
\value{
Named vector of generation numbers
}
\description{
Return number of generations for individuals
}
