% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigree_stats.R
\name{gen_numbers_living}
\alias{gen_numbers_living}
\title{Get generation numbers for living individuals}
\usage{
gen_numbers_living(pedigree.living, studbook)
}
\arguments{
\item{pedigree.living}{Pedigree object}

\item{studbook}{Studbook tibble}
}
\value{
Named vector of generations for living IDs
}
\description{
Get generation numbers for living individuals
}
