% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigree_stats.R
\name{kin_matrix_living}
\alias{kin_matrix_living}
\title{Subset kinship matrix to living individuals}
\usage{
kin_matrix_living(pedigree.living, studbook)
}
\arguments{
\item{pedigree.living}{A pedigree object}

\item{studbook}{Studbook tibble}
}
\value{
Sub-matrix of pairwise kinship among living individuals
}
\description{
Subset kinship matrix to living individuals
}
