% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigree_stats.R
\name{kinship_summary}
\alias{kinship_summary}
\title{Summarize pedigree-wide statistics}
\usage{
kinship_summary(pedigree, pedigree.living, studbook)
}
\arguments{
\item{pedigree}{Full pedigree object}

\item{pedigree.living}{Living-only pedigree object}

\item{studbook}{Studbook tibble}
}
\value{
A tibble of demographic/genetic statistics
}
\description{
Summarize pedigree-wide statistics
}
