% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{lighten_plotly_pal}
\alias{lighten_plotly_pal}
\title{Lighten color palette for plotly usage}
\usage{
lighten_plotly_pal(palette, hex = "33")
}
\arguments{
\item{palette}{A named vector of colors}

\item{hex}{A two-character hex code representing transparency}
}
\value{
A named vector of lightened colors
}
\description{
Lighten color palette for plotly usage
}
