% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactable_col_lists.R
\name{living.groups}
\alias{living.groups}
\title{Column groupings for living summary table}
\usage{
living.groups()
}
\value{
A named list of column groups
}
\description{
Column groupings for living summary table
}
