% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{make_cohorts}
\alias{make_cohorts}
\title{Generate cohort-structured data}
\usage{
make_cohorts(df, minYear, maxYear, span, maxAge, include_sex = TRUE)
}
\arguments{
\item{df}{A tibble of birth records}

\item{minYear}{Start year of cohorts}

\item{maxYear}{End year}

\item{span}{Years per cohort}

\item{maxAge}{Max age to include}

\item{include_sex}{Whether to include sex as a grouping var}
}
\value{
A joined and restructured tibble
}
\description{
Generate cohort-structured data
}
