% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{matrix.heatmap}
\alias{matrix.heatmap}
\title{Generate a kinship or relatedness heatmap using heatmaply}
\usage{
matrix.heatmap(matrix, title, key.title, xlab = "Females", ylab = "Males")
}
\arguments{
\item{matrix}{A numeric matrix of pairwise values (e.g., kinship)}

\item{title}{Title for the heatmap}

\item{key.title}{Title for the color bar legend}

\item{xlab}{Optional x-axis label}

\item{ylab}{Optional y-axis label}
}
\value{
A heatmaply widget
}
\description{
Generate a kinship or relatedness heatmap using heatmaply
}
