% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/life_tables.R
\name{MLE_age1}
\alias{MLE_age1}
\title{Estimate age at 50\% survivorship (MLE age 1+)}
\usage{
MLE_age1(lx, age)
}
\arguments{
\item{lx}{A vector of Lx values}

\item{age}{A vector of corresponding ages}
}
\value{
A single numeric value representing the interpolated MLE age at lx1 = 0.5
}
\description{
This function returns the age at which the first-life cohort (lx1)
reaches 0.5 using linear interpolation between the nearest age classes.
}
\keyword{internal}
