% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{nest_timeline}
\alias{nest_timeline}
\title{Nest timeline data by group}
\usage{
nest_timeline(timeline, groupBy = NULL)
}
\arguments{
\item{timeline}{Timeline tibble}

\item{groupBy}{Optional grouping variable (e.g., "Location", "Age", "Sex")}
}
\value{
Nested list by Date
}
\description{
Nest timeline data by group
}
