% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{palette_zoolabs}
\alias{palette_zoolabs}
\title{Define default zoolabs color palette}
\usage{
palette_zoolabs()
}
\value{
A named list of core colors for use in plots, pedigrees, and tables
}
\description{
Define default zoolabs color palette
}
