% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigree_network.R
\name{ped_network}
\alias{ped_network}
\title{Build visNetwork visualization of a pedigree}
\usage{
ped_network(pedigree, palette, studbook)
}
\arguments{
\item{pedigree}{A pedigree object}

\item{palette}{A color palette in list format}

\item{studbook}{A studbook tibble}
}
\value{
A visNetwork plot object
}
\description{
Build visNetwork visualization of a pedigree
}
