% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigree_network.R
\name{pedigree_levels}
\alias{pedigree_levels}
\title{Assign generation levels to individuals in a pedigree}
\usage{
pedigree_levels(pedigree, studbook)
}
\arguments{
\item{pedigree}{A pedigree object from \code{pedtools::ped()}}

\item{studbook}{A studbook tibble}
}
\value{
A tibble of individual IDs and generation levels
}
\description{
Assign generation levels to individuals in a pedigree
}
