% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigree_network.R
\name{pedigree_pairs}
\alias{pedigree_pairs}
\title{Extract unique mating pairs from a pedigree}
\usage{
pedigree_pairs(pedigree, studbook)
}
\arguments{
\item{pedigree}{A pedigree object}

\item{studbook}{A studbook tibble}
}
\value{
A tibble of unique mating pairs with location info
}
\description{
Extract unique mating pairs from a pedigree
}
