% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_census}
\alias{plot_census}
\title{Plot population trends using census data}
\usage{
plot_census(census_df, palette, title)
}
\arguments{
\item{census_df}{A tibble with columns Date, Females, Males, Unidentified}

\item{palette}{A named list of colors for "f", "m", and "u"}

\item{title}{The title of the plot}
}
\value{
A plotly plot object
}
\description{
Plot population trends using census data
}
