% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_lambda}
\alias{plot_lambda}
\title{Plot lambda growth rates by cohort}
\usage{
plot_lambda(life_table, title)
}
\arguments{
\item{life_table}{A tibble with lambda values and Sex, CohortLabel columns}

\item{title}{Plot title}
}
\value{
A plotly object
}
\description{
Plot lambda growth rates by cohort
}
