% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_ped_series}
\alias{plot_ped_series}
\title{Plot a series of pedigree objects using pedtools with custom formatting}
\usage{
plot_ped_series(ped.list, palette, studbook)
}
\arguments{
\item{ped.list}{A list of pedigree objects (from \code{pedtools::ped()})}

\item{palette}{A named list of color values}

\item{studbook}{A tibble with \code{ID}, \code{Sex}, \code{Status}, and other metadata}

\item{...}{Additional arguments passed to \code{pedtools::plot()}}
}
\value{
A series of pedigree plots (base graphics)
}
\description{
Plot a series of pedigree objects using pedtools with custom formatting
}
