% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigree_stats.R
\name{rel_founder_contributions}
\alias{rel_founder_contributions}
\title{Calculate relative founder contributions}
\usage{
rel_founder_contributions(studbook, pedigree.living)
}
\arguments{
\item{studbook}{Studbook tibble}

\item{pedigree.living}{Pedigree for living individuals}
}
\value{
A numeric vector of relative contributions summing to 1
}
\description{
Calculate relative founder contributions
}
