% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{set_ped_fills}
\alias{set_ped_fills}
\title{Generate fill color mapping for pedigree plots}
\usage{
set_ped_fills(pedigree, studbook)
}
\arguments{
\item{pedigree}{A pedigree object created by \code{pedtools::ped()}}

\item{studbook}{A tibble of studbook data}
}
\value{
A named list mapping individuals to colors by sex and status
}
\description{
Generate fill color mapping for pedigree plots
}
