% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{living}
\alias{living}
\alias{deceased}
\alias{living.males}
\alias{living.females}
\title{Return vector of living or deceased IDs}
\usage{
living(studbook)

deceased(studbook)

living.males(studbook)

living.females(studbook)
}
\arguments{
\item{studbook}{Studbook data}
}
\description{
Return vector of living or deceased IDs
}
