% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{subset_matrix_living}
\alias{subset_matrix_living}
\title{Subset a full kinship matrix to include only living males vs. living females}
\usage{
subset_matrix_living(matrix, studbook)
}
\arguments{
\item{matrix}{A symmetric kinship matrix}

\item{studbook}{A tibble with metadata for all individuals}
}
\value{
A matrix with rows = males, cols = females, both alive
}
\description{
Subset a full kinship matrix to include only living males vs. living females
}
