% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{palette_zoolabs}
\alias{palette_zoolabs}
\title{Define standardized color set for zoo biology plots}
\usage{
palette_zoolabs()
}
\value{
A named list of color hex codes
}
\description{
Define standardized color set for zoo biology plots
}
