% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigree_network.R
\name{pedigree_nodes}
\alias{pedigree_nodes}
\title{Generate node list for pedigree visNetwork plot}
\usage{
pedigree_nodes(pedigree, studbook)
}
\arguments{
\item{pedigree}{A pedigree object}

\item{studbook}{A studbook tibble}
}
\value{
A tibble of nodes with visNetwork styling attributes
}
\description{
Generate node list for pedigree visNetwork plot
}
