% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigree_network.R
\name{subset_network_btp}
\alias{subset_network_btp}
\title{Subset pedigree network to visualize match lineage}
\usage{
subset_network_btp(pedigree.living, palette, pair, studbook)
}
\arguments{
\item{pedigree.living}{A pedigree object for the living population}

\item{palette}{A color palette in list format}

\item{pair}{A character vector of two individual IDs (male, female)}

\item{studbook}{A tibble containing metadata for the full population}
}
\value{
A visNetwork object showing related individuals and lineage paths
}
\description{
Subset pedigree network to visualize match lineage
}
